+++ docs/news.rst
@@ -3,6 +3,7 @@ Release Notes
 
 **1.3.0 (unreleased)**
 
+* Moved retrieval of Firefox path to ``firefox_path`` fixture.
 * Added driver and sensitive URL to report header.
 * Moved base URL implementation to the pytest-base-url plugin.