+++ docs/news.rst
@@ -6,6 +6,7 @@ Release Notes
 * Moved retrieval of Firefox path to ``firefox_path`` fixture.
 * Added driver and sensitive URL to report header.
 * Moved base URL implementation to the pytest-base-url plugin.
+* Moved retrieval of Firefox path to `firefox_path` fixture
 
 **1.2.1 (2016-02-25)**